Applied several fixes for Clang 5.
Clang is a little more peculiar about the warnings for unused function
arguments, etc.  Also disabled the tautological compare warning since
it was comparing with a variable that happened to be 0, but that requires
knowledge beyond that which is available to the function.

+++ include/terminalpp/glyph.hpp
@@ -79,7 +79,7 @@ public:
     {
         for (size_t index = 0; index < sizeof(ucharacter_); ++index)
         {
-            ucharacter_[index] = ustr[index];
+            ucharacter_[index] = static_cast<byte>(ustr[index]);
 
             if (!(ucharacter_[index] & 0x80))
             {
